pyproject.toml
[project]
name = "money-collect-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.0",
    "django>=5.2.9",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "flake8>=7.3.0",
    "pillow>=12.0.0",
    "psycopg>=3.3.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
]
